Moving components to separate file
